feat(router): adds wildcard route matching
This expands the route DSL to use two
new syntaxes.

a single `*` is used to create a wildcard
that matches any valid URL string but does not
expand across multiple segments.

a double `**` is used to create a wildcard
that matches any valid URL string and does
expand across multiple segments.

Fixes #5