fix(*): don't force StaticFileHandler to be last middleware
Before this change the StaticFileHandler middleware
directly yielded the 404 error which effectively
means it *must* be the last middleware in the stack.

With this change the StaticFileHandler middleware
can be anywhere in the stack. 404 erros are caught
by forcing a last Middleware handler that catches
all unprocessed requests and yield the 404 error.

Fixes #61