fix(request): request.param(_) now returns an Option
A user may wish to use the same handler for multiple routes,
if the behaviour of such a handler should differ based on the
available params then the handler it will panic at runtime as
there is no convenient way to check for the presence of the
param before accessing it.

For broken code:
If you have `foo.param("bar")` then you'll need to `unwrap` the
Option, e.g. `foo.param("bar").unwrap()`.

[breaking-change]