feat(server): add some shared data across all requests
This can be used to add some compile time dependencies for
Middleware and Plugins.

To fix code broken by this, you will need to introduce a type
parameter for Request, Response, MiddlewareResult and NickelError.

```
// e.g. This
fn foo<'a>(&mut Request, Response<'a>) -> MiddlewareResult<'a>

// Should become:
fn foo<'a, D>(&mut Request<D>, Response<'a, D>) -> MiddlewareResult<'a, D>
```

You can add bounds to `D` in the above to place compile-time restrictions on
the server data (can be used for configuration).

[breaking-change]