refactor(*): Request is now accessible through a Response instance
This is motivated to allow writing Plugins which are able to access
both the current Response and Request without having to rely on odd
syntax such as `(foo, bar).baz()`.

BREAKING CHANGE: This breaks a *lot* of nickel usage, checking the diff
from the examples should help to migrate code relying on the old structure.