fix(error): introduce IntoError trait to help with coherence problems
The previous reliance on the From and Into traits lead to coherence
violations for users who may want to implement anything more than a
simple mapping to a String or StatusCode.

BREAKING CHANGE:
The error_handling example and the IntoError trait documentation
have a number of different examples of trait implementation, for both
foreign types and crate local types. Hopefully they should cover most
of the usage which may be broken from this change.