fix(examples): Fix examples triggering issue #399
Fix the examples that were returning MiddlewareResults in the
middlware! macro. A couple could be changed to return a Responder, but
most needed to be reimplemented as seperate functions.

Updated the middleware! macro documentaion to note the limitations
from issues #399 and #389.

Disable ssl testing in travis, which was broken in commit
8d5a7d0515dab850d88249f392db8db463437068. Issue #415 to track
our SSL plan.